intercompany_requests2/requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
django-environ==0.11.2
pillow==10.1.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-htmx==1.17.2
whitenoise==6.6.0
gunicorn==21.2.0
django-allauth==0.60.1
python-decouple==3.8